% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_heading.R
\name{gps_heading}
\alias{gps_heading}
\title{Derive heading based on consecutive geographic locations}
\usage{
gps_heading(lon, lat, degree = TRUE)
}
\description{
Derive heading based on consecutive geographic locations
}
