% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_speed_mph.R
\name{gps_speed_mph}
\alias{gps_speed_mph}
\title{Derive speed based on geographic location}
\usage{
gps_speed_mph(lon, lat, time)
}
\arguments{
\item{lon}{vector of geographic longitudes}

\item{lat}{vector of geographic latitudes}

\item{time}{vector of time stamps}
}
\value{
vector of speed in miles per hour (mph)
}
\description{
Time and distance between consecutive geographic locations are used to
derive the speed at which a trace is traversed.
}
