% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gpx.R
\name{read_gpx}
\alias{read_gpx}
\title{Read OSM traces from a gpx file}
\usage{
read_gpx(file)
}
\arguments{
\item{file}{file path to a gpx file}
}
\value{
data frame of the traces
}
\description{
The OpenStreetMap API returns traces in form of paged gpx (v1.0) files.
This function should not be used as a stand-alone gpx reader - it is meant only
to read the downloaded trace files.
}
\examples{
# example code
library(ggplot2)
read_gpx(file.path(system.file(package = "osmtraces"),"ames/page-0.gpx")) |>
  tidyr::unnest(trkseg) |>
  ggplot(aes(x = lon, y = lat)) + geom_point()
}
