% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentize.R
\name{segmentize}
\alias{segmentize}
\title{Turn a track into multiple segments}
\usage{
segmentize(time, seconds = 5)
}
\arguments{
\item{time}{vector of time stamps}

\item{seconds}{positive value indicating the time allowed between successive time
points before a new segment is started.}
}
\value{
integer vector of segment ids
}
\description{
If the time between successive time points exceeds the interval specified
in seconds, it is assumed hat a new segment has started.
}
