% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_to_df.R
\name{trace_to_df}
\alias{trace_to_df}
\title{Convert a trace into a data frame}
\usage{
trace_to_df(trk)
}
\arguments{
\item{trk}{xml node with elements and attributes}
}
\value{
single row dataframe with a nested dataframe of the trace
in form of a tibble with variables latitude, longitude, and time.
}
\description{
Turns a GPX defined track from a format as xml node into a data frame.
All elements are turned into variables, the track is a single-entry
list variable with a data frame of time-stamped geographic locations.
This function is called as part of \code{read_gpx}. There should be no need to call
this function by itself.
}
